- updated contributors & removed bad contributor code \n - moved last 2 challenges one number down \n - fixed test
