+++ src/main/java/org/owasp/wrongsecrets/challenges/docker/Challenge70Controller.java
@@ -1,10 +1,9 @@
 package org.owasp.wrongsecrets.challenges.docker;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.http.ContentDisposition;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -12,33 +11,31 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * Hosts the Cursor skill of challenge 70 straight from the resource folder, so participants can
- * read the skill the same way an agent would.
+ * Hosts the Claude skill of challenge 71. The exported skill bundle lives in the resource folder
+ * and is served directly so participants download the same file that was exported from Claude.
  */
 @Slf4j
 @RestController
 public class Challenge70Controller {
 
-  private static final MediaType MARKDOWN =
-      new MediaType("text", "markdown", StandardCharsets.UTF_8);
+  static final String SKILL_ROOT = "challenges/challenge-70/claude-skill/";
+  private static final String BUNDLE_NAME = "incident-reporter.zip";
+  private static final MediaType ZIP = new MediaType("application", "zip");
 
-  private final Resource skillFile;
-
-  public Challenge70Controller(
-      @Value("classpath:challenges/challenge-70/cursor-skill/deploy-preview/SKILL.md")
-          Resource skillFile) {
-    this.skillFile = skillFile;
-  }
-
-  /** Returns the raw {@code SKILL.md} of the {@code deploy-preview} Cursor skill. */
-  @GetMapping("/skills/cursor/deploy-preview/SKILL.md")
-  public ResponseEntity<String> cursorSkill() {
+  /** Returns the exported {@code incident-reporter} Claude skill as a downloadable zip bundle. */
+  @GetMapping("/skills/claude/incident-reporter.zip")
+  public ResponseEntity<byte[]> claudeSkillBundle() {
     try {
+      var resource = new ClassPathResource(SKILL_ROOT + "incident-reporter.skill");
       return ResponseEntity.ok()
-          .contentType(MARKDOWN)
-          .body(skillFile.getContentAsString(StandardCharsets.UTF_8));
+          .contentType(ZIP)
+          .headers(
+              headers ->
+                  headers.setContentDisposition(
+                      ContentDisposition.attachment().filename(BUNDLE_NAME).build()))
+          .body(resource.getContentAsByteArray());
     } catch (IOException e) {
-      log.warn("Unable to serve the Cursor skill of challenge 70", e);
+      log.warn("Unable to serve the Claude skill of challenge 71", e);
       return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
   }
