+++ src/main/resources/challenges/challenge-69/cursor-skill/deploy-preview/SKILL.md
