+++ src/main/resources/challenges/challenge-70/challenge-70.snippet
@@ -1,38 +1,18 @@
-<div id="cursor-skill-container" style="border: 1px solid #ccc; border-radius: 8px; padding: 20px; margin: 20px; background-color: #f9f9f9;">
-    <h4>📄 Cursor skill: <code>deploy-preview</code></h4>
-    <p>This application ships a Cursor skill in its resource folder. The backend hosts it at <code>/skills/cursor/deploy-preview/SKILL.md</code>, which is exactly what an agent would read when the skill is installed.</p>
+<div id="claude-skill-container" style="border: 1px solid #ccc; border-radius: 8px; padding: 20px; margin: 20px; background-color: #f9f9f9;">
+    <h4>📦 Claude skill bundle: <code>incident-reporter</code></h4>
+    <p>This application ships an exported Claude skill bundle. The backend serves the exported <code>.skill</code> file as a zip download at <code>/skills/claude/incident-reporter.zip</code>.</p>
 
     <div class="skill-warning" style="border: 1px solid #ffeaa7; border-radius: 6px; padding: 15px; margin: 15px 0;">
-        <p>Fetch the skill from the command line:</p>
-        <pre class="skill-code"
-             style="padding: 10px; border-radius: 4px; overflow-x: auto; font-size: 13px;"
-             th:text="'curl -s ' + ${#httpServletRequest.scheme} + '://' + ${#httpServletRequest.serverName} + (${#httpServletRequest.serverPort == 80 || #httpServletRequest.serverPort == 443 ? '' : ':' + #httpServletRequest.serverPort}) + '/skills/cursor/deploy-preview/SKILL.md'">curl -s http://localhost:8080/skills/cursor/deploy-preview/SKILL.md</pre>
+        <p>Step 1 — download and unpack the bundle:</p>
+        <pre class="skill-code" style="padding: 10px; border-radius: 4px; overflow-x: auto; font-size: 13px;"
+             th:text="'curl -sO ' + ${httpServletRequest.scheme} + '://' + ${httpServletRequest.serverName} + (${httpServletRequest.serverPort == 80 || httpServletRequest.serverPort == 443 ? '' : ':' + httpServletRequest.serverPort}) + '/skills/claude/incident-reporter.zip'">curl -sO http://localhost:8080/skills/claude/incident-reporter.zip</pre>
+        <pre class="skill-code" style="padding: 10px; border-radius: 4px; overflow-x: auto; font-size: 13px;">unzip incident-reporter.zip
+find incident-reporter -type f</pre>
+        <a class="btn btn-warning btn-sm skill-btn" href="/skills/claude/incident-reporter.zip">⬇ Download incident-reporter.zip</a>
 
-        <p style="margin-top: 12px;">…or read it right here:</p>
-        <button class="btn btn-warning btn-sm skill-btn" onclick="loadCursorSkill()">▶ Show SKILL.md</button>
-        <a class="btn btn-secondary btn-sm skill-btn" href="/skills/cursor/deploy-preview/SKILL.md" target="_blank" rel="noopener">Open in new tab</a>
-        <pre id="cursor-skill-output" class="skill-output" style="display:none; padding: 10px; border-radius: 4px; overflow-x: auto; font-size: 12px; margin-top: 8px; white-space: pre-wrap; word-break: break-word;"></pre>
+        <p style="margin-top: 12px;">Step 2 — the <code>SKILL.md</code> is clean. Look at what it tells the agent to run:</p>
+        <pre class="skill-code" style="padding: 10px; border-radius: 4px; overflow-x: auto; font-size: 13px;">grep -r TOKEN incident-reporter/</pre>
 
-        <p style="margin-top: 12px;"><small class="skill-tip">💡 The author inlined the shared staging deploy token so the skill would be "self-contained". Submit that token.</small></p>
+        <p style="margin-top: 12px;"><small class="skill-tip">💡 The token you find is not the answer yet — the author "hid" it from the secret scanner. Submit the value the uploader actually authenticates with.</small></p>
     </div>
 </div>
-
-<style>
-.skill-output {
-    background-color: #f8f9fa !important;
-    color: #212529 !important;
-    border: 1px solid #dee2e6 !important;
-}
-</style>
-
-<script>
-function loadCursorSkill() {
-    const out = document.getElementById('cursor-skill-output');
-    out.style.display = 'block';
-    out.textContent = 'Loading…';
-    fetch('/skills/cursor/deploy-preview/SKILL.md')
-        .then(function(r) { return r.text(); })
-        .then(function(text) { out.textContent = text; })
-        .catch(function(err) { out.textContent = 'Failed to load the skill file: ' + (err.message || err); });
-}
-</script>
