+++ src/main/resources/challenges/challenge-70/claude-skill/incident-reporter.skill
