+++ src/main/resources/explanations/challenge70_reason.adoc
@@ -1,24 +1,26 @@
-*Why you should never put a secret in an agent skill*
+*Why a skill bundle needs the same review as your source code*
 
-A skill is built to be shared. That is the whole point of it: you write the workflow once and every developer, agent and CI job reuses it. This means a secret in a skill has the worst possible properties:
+A skill bundle is shipped software. It contains executable scripts, it runs on developer machines and in CI, and it is installed by people who did not write it. Treating it as "just some markdown" is how secrets end up being distributed at scale.
 
-- It is copied to every machine that installs the skill, including personal laptops.
-- It ends up in the repository, so it is in the git history forever, even after you "remove" it, forcing you to rotate the secret.
-- It is published to skill marketplaces and registries, where it is indexed and searchable.
-- It is read by the agent and can therefore be echoed into chat transcripts, logs and telemetry belonging to third parties.
-- It is a *shared* credential, so it is almost never rotated and almost never scoped down.
+Two failures stack up in this challenge:
+
+- The token is committed inside the bundle, so everyone who installs the skill receives a working shared credential.
+- When sharing the skill inside Claude, everyone with access to the skill, will have access to the secret.
+- The token is base64 encoded, which is *encoding*, not encryption. It stops a naive secret scanner from firing, and it stops nobody else. Anyone with the bundle can decode it in one command.
+
+The second point is the more dangerous of the two, because it converts a detectable problem into an undetectable one. The scanner goes quiet, the finding disappears from the backlog, and the credential stays valid for years.
 
 ----
 What to do instead:
 
-- Keep credentials out of the skill and let the skill read them from the environment, for example `$STAGING_DEPLOY_TOKEN`.
-- Document *which* variable is required and *where* to obtain it, never the value itself.
-- Give every consumer their own short-lived, least-privilege credential instead of one shared team token.
-- Scan skill files with a secret scanner in pre-commit and CI, just like you scan source code.
-- Treat a secret that ever appeared in a skill as compromised and rotate it.
+- Read credentials from the environment or a secret manager at runtime, and fail loudly with a clear message when they are absent.
+- Never "fix" a secret scanner finding by encoding, splitting or obfuscating the value. Fix it by removing the secret and rotating it.
+- Review every file in a skill bundle, not just the `SKILL.md`. Scripts, reference documents and sample configuration are all part of the attack surface.
+- Verify the bundle in CI: unpack it and run your secret scanner over the extracted contents, including a base64-aware ruleset.
+- Publish bundles from a build pipeline that has no access to production credentials in the first place.
 ----
 
 [NOTE]
 ====
-Skills, rules, prompts, MCP server configurations and agent instruction files are all code as far as your secret management is concerned. If you would not hardcode a token in a `.java` file, do not hardcode it in a `SKILL.md` either.
+Encoding is not a security control. If the only thing standing between an attacker and your credential is a `base64 -d`, the credential is public.
 ====
