--- src/main/resources/explanations/challenge71.adoc
@@ -1,16 +1,0 @@
-=== Challenge 71: Find the Secret in the Claude Skill Bundle
-
-Claude skills are distributed as exported `.skill` bundles: a zip archive containing a folder with a `SKILL.md` that describes when to use the skill, plus any scripts and reference material the skill needs. The whole bundle is unpacked on the machine that installs it.
-
-That makes a skill bundle a great hiding place, because reviewers tend to read the `SKILL.md` and stop there. The interesting part is usually in the files next to it.
-
-This application ships a Claude skill called `incident-reporter`. The exported skill bundle lives in the resource folder and is served as a download at link:/skills/claude/incident-reporter.zip[`/skills/claude/incident-reporter.zip`].
-
-Download the bundle, unpack it, and work out the token the uploader authenticates with.
-
-The skill source is also available on GitHub at link:https://github.com/OWASP/wrongsecrets/blob/master/src/main/resources/challenges/challenge-71/claude-skill/incident-reporter.skill[`src/main/resources/challenges/challenge-71/claude-skill/incident-reporter.skill`].
-
-[NOTE]
-====
-The `SKILL.md` is clean. Reading only the entry point of a skill is not a review.
-====
