+++ src/main/resources/wrong-secrets-configuration.yaml
@@ -1013,7 +1013,7 @@ configurations:
       category: *bin
       ctf:
         enabled: true
-        
+
     - name: Challenge 68
       short-name: "challenge-68"
       sources:
@@ -1026,7 +1026,20 @@ configurations:
       category: *ai
       ctf:
         enabled: true
-        
+
+    - name: Challenge 69
+      short-name: "challenge-69"
+      sources:
+        - class-name: "org.owasp.wrongsecrets.challenges.docker.Challenge69"
+          explanation: "explanations/challenge69.adoc"
+          hint: "explanations/challenge69_hint.adoc"
+          reason: "explanations/challenge69_reason.adoc"
+          ui-snippet: "challenges/challenge-69/challenge-69.snippet"
+          environments: *all_envs
+      difficulty: *easy
+      category: *ai
+      ctf:
+        enabled: true
     - name: Challenge 70
       short-name: "challenge-70"
       sources:
@@ -1035,19 +1048,6 @@ configurations:
           hint: "explanations/challenge70_hint.adoc"
           reason: "explanations/challenge70_reason.adoc"
           ui-snippet: "challenges/challenge-70/challenge-70.snippet"
-          environments: *all_envs
-      difficulty: *easy
-      category: *ai
-      ctf:
-        enabled: true
-    - name: Challenge 71
-      short-name: "challenge-71"
-      sources:
-        - class-name: "org.owasp.wrongsecrets.challenges.docker.Challenge71"
-          explanation: "explanations/challenge71.adoc"
-          hint: "explanations/challenge71_hint.adoc"
-          reason: "explanations/challenge71_reason.adoc"
-          ui-snippet: "challenges/challenge-71/challenge-71.snippet"
           environments: *all_envs
       difficulty: *normal
       category: *ai
