+++ src/test/java/org/owasp/wrongsecrets/challenges/docker/Challenge70ControllerTest.java
@@ -2,40 +2,72 @@ package org.owasp.wrongsecrets.challenges.docker;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.zip.ZipInputStream;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpStatus;
 
 class Challenge70ControllerTest {
 
-  private static final String SKILL_LOCATION =
-      "challenges/challenge-70/cursor-skill/deploy-preview/SKILL.md";
+  private static final String SKILL_MD = "incident-reporter/SKILL.md";
+  private static final String UPLOADER = "incident-reporter/scripts/upload_report.py";
+  private static final String RUNBOOK = "incident-reporter/references/runbook.md";
+
+  private static Map<String, String> unzip(byte[] bundle) throws IOException {
+    var entries = new HashMap<String, String>();
+    try (var zip = new ZipInputStream(new ByteArrayInputStream(bundle))) {
+      for (var entry = zip.getNextEntry(); entry != null; entry = zip.getNextEntry()) {
+        entries.put(entry.getName(), new String(zip.readAllBytes(), StandardCharsets.UTF_8));
+      }
+    }
+    return entries;
+  }
 
   @Test
-  void shouldServeTheCursorSkillAsMarkdown() {
-    var controller = new Challenge70Controller(new ClassPathResource(SKILL_LOCATION));
+  void bundleShouldContainAllSkillFilesRelativeToTheSkillFolder() throws IOException {
+    var response = new Challenge70Controller().claudeSkillBundle();
 
-    var response = controller.cursorSkill();
+    var entries = unzip(response.getBody());
+    assertThat(entries).containsKey(SKILL_MD);
+    assertThat(entries).containsKey(UPLOADER);
+    assertThat(entries).containsKey(RUNBOOK);
+  }
+
+  @Test
+  void bundledSkillFileShouldNotContainTheSecret() throws IOException {
+    var response = new Challenge70Controller().claudeSkillBundle();
+
+    var entries = unzip(response.getBody());
+    assertThat(entries.get(SKILL_MD)).contains("name: incident-reporter");
+    assertThat(entries.get(SKILL_MD)).doesNotContain("UPLOAD_TOKEN_B64");
+  }
+
+  @Test
+  void bundledUploaderShouldCarryTheEncodedTokenAndNotThePlaintextOne() throws IOException {
+    var response = new Challenge70Controller().claudeSkillBundle();
+
+    var entries = unzip(response.getBody());
+    var challenge =
+        new Challenge70(
+            new ClassPathResource(Challenge70Controller.SKILL_ROOT + "incident-reporter.skill"));
+
+    assertThat(entries.get(UPLOADER)).contains("UPLOAD_TOKEN_B64 = \"");
+    assertThat(entries.get(UPLOADER)).doesNotContain(challenge.spoiler().solution());
+  }
+
+  @Test
+  void shouldServeTheBundleAsAZipDownload() {
+    var response = new Challenge70Controller().claudeSkillBundle();
 
     assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-    assertThat(response.getHeaders().getContentType()).hasToString("text/markdown;charset=UTF-8");
-    assertThat(response.getBody()).contains("name: deploy-preview", "STAGING_DEPLOY_TOKEN=\"");
-  }
-
-  @Test
-  void servedSkillShouldContainTheAnswerOfTheChallenge() {
-    var controller = new Challenge70Controller(new ClassPathResource(SKILL_LOCATION));
-    var challenge = new Challenge70(new ClassPathResource(SKILL_LOCATION));
-
-    assertThat(controller.cursorSkill().getBody()).contains(challenge.spoiler().solution());
-  }
-
-  @Test
-  void shouldReturnServerErrorWhenTheSkillIsMissing() {
-    var controller =
-        new Challenge70Controller(new ClassPathResource("challenges/challenge-70/missing.md"));
-
-    assertThat(controller.cursorSkill().getStatusCode())
-        .isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
+    assertThat(response.getHeaders().getContentType()).hasToString("application/zip");
+    assertThat(response.getHeaders().getContentDisposition().getFilename())
+        .isEqualTo("incident-reporter.zip");
+    assertThat(response.getBody()).isNotEmpty();
   }
 }
